+++ shuup_tests/admin/test_refund_creator.py
@@ -10,7 +10,6 @@ import pytest
 
 from bs4 import BeautifulSoup
 from django.test import override_settings
-
 from shuup.admin.modules.orders.views.refund import (
     OrderCreateFullRefundView, OrderCreateRefundView
 )
@@ -58,6 +57,75 @@ def test_create_refund_view(rf, admin_user):
     refund_line = order.lines.filter(type=OrderLineType.REFUND).last()
     assert refund_line
     assert refund_line.taxful_price == -product_line.taxful_price
+
+
+@pytest.mark.django_db
+def test_create_amount_refund(rf, admin_user):
+    shop = get_default_shop()
+    supplier = get_default_supplier()
+    product = create_product(sku="test-sku", shop=shop, supplier=supplier, default_price=3.33)
+    order = create_order_with_product(product, supplier, quantity=1, taxless_base_unit_price=1, shop=shop)
+    order.cache_prices()
+    order.save()
+
+    assert not order.has_refunds()
+    assert len(order.lines.all()) == 1
+
+    product_line = order.lines.first()
+
+    data = {
+        "form-0-line_number": "amount",
+        "form-0-text": "Test refund",
+        "form-0-amount": 1,
+        "form-0-supplier": supplier.id,
+        "form-INITIAL_FORMS": 0,
+        "form-MAX_NUM_FORMS": 1000,
+        "form-TOTAL_FORMS": 1,
+        "form-MIN_NUM_FORMS": 0,
+    }
+
+    request = apply_request_middleware(rf.post("/", data=data), user=admin_user)
+    view = OrderCreateRefundView.as_view()
+    response = view(request, pk=order.pk)
+    assert response.status_code == 302
+    assert order.has_refunds()
+
+    assert len(order.lines.all()) == 2
+    refund_line = order.lines.filter(type=OrderLineType.REFUND).last()
+    assert refund_line
+    assert refund_line.taxful_price == -product_line.taxful_price
+
+
+@pytest.mark.django_db
+def test_create_to_big_amount_refund(rf, admin_user):
+    shop = get_default_shop()
+    supplier = get_default_supplier()
+    product = create_product(sku="test-sku", shop=shop, supplier=supplier, default_price=3.33)
+    order = create_order_with_product(product, supplier, quantity=1, taxless_base_unit_price=1, shop=shop)
+    order.cache_prices()
+    order.save()
+
+    assert not order.has_refunds()
+    assert len(order.lines.all()) == 1
+
+    product_line = order.lines.first()
+
+    data = {
+        "form-0-line_number": "amount",
+        "form-0-text": "Test refund",
+        "form-0-amount": 10,
+        "form-0-supplier": supplier.id,
+        "form-INITIAL_FORMS": 0,
+        "form-MAX_NUM_FORMS": 1000,
+        "form-TOTAL_FORMS": 1,
+        "form-MIN_NUM_FORMS": 0,
+    }
+
+    request = apply_request_middleware(rf.post("/", data=data), user=admin_user)
+    view = OrderCreateRefundView.as_view()
+    response = view(request, pk=order.pk)
+    assert response.status_code == 200
+    assert not order.has_refunds()
 
 
 @pytest.mark.django_db
@@ -113,6 +181,28 @@ def test_arbitrary_refund_availability(rf, admin_user):
     assert refund_option_str in get_refund_view_content()
     with override_settings(SHUUP_ALLOW_ARBITRARY_REFUNDS=False):
         assert refund_option_str not in get_refund_view_content()
+
+
+@pytest.mark.django_db
+def test_supplier_select_for_admin(rf, admin_user):
+    shop = get_default_shop()
+    supplier = get_default_supplier()
+    product = create_product(sku="test-sku", shop=shop, supplier=supplier, default_price=3.33)
+    order = create_order_with_product(product, supplier, quantity=1, taxless_base_unit_price=1, shop=shop)
+    order.cache_prices()
+    order.save()
+
+    assert not order.has_refunds()
+    assert len(order.lines.all()) == 1
+
+    def get_refund_view_content():
+        request = apply_request_middleware(rf.get("/"), user=admin_user)
+        view = OrderCreateRefundView.as_view()
+        response = view(request, pk=order.pk)
+        return force_text(response.render().content)
+
+    supplier_option_str = '$refund.find("select[id$=\'supplier\']"), true);'
+    assert supplier_option_str in get_refund_view_content()
 
 
 @pytest.mark.django_db